requirements.txt
torch==2.1.0
kaggle~=1.5.16
joblib~=1.3.2
scikit-learn~=1.3.1
torchvision~=0.16.0
Pillow~=10.1.0
tqdm~=4.66.1
numpy~=1.26.1
pandas~=2.1.1